dev: change payment schema

+++ database/migrations/2025_12_15_161417_create_cash_transactions_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::create('cash_transactions', function (Blueprint $table) {
-            $table->unsignedBigInteger('id')->primary();
+            $table->increments('id')->primary();
             $table->string('reference_type');
             $table->unsignedBigInteger('reference_id');
             $table->date('transaction_date');
